requirements.txt
aiogram==3.10.0
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.0
pyyaml==6.0.2
aiohttp==3.10.5
